Make into a package
Packages are easier to install than raw scripts. Just add to your
manifest and you're off!

Move code into a folder and add an asmdef.

Add metas (and remove meta from gitignore).
