+++ Editor/FindReferencesInProject2.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 28b93a2d809b7f0439ed1828f60ab705
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
