+++ Editor/findrefs.Editor.asmdef
@@ -1,0 +1,10 @@
+{
+    "name": "findrefs.Editor",
+    "references": [
+    ],
+    "optionalUnityReferences": [
+    ],
+    "includePlatforms": [
+        "Editor"
+    ]
+}
