+++ Editor/ignore.txt
@@ -1,0 +1,72 @@
+# binary files
+*.FBX
+*.TGA
+*.a
+*.aar
+*.ab
+*.bmp
+*.bundle
+*.bytes
+*.chm
+*.chw
+*.cubemap
+*.dds
+*.dll
+*.exe
+*.fbx
+*.gif
+*.jpg
+*.mdb
+*.mel
+*.mp3
+*.otf
+*.pdf
+*.png
+*.psd
+*.swatch
+*.tga
+*.tif
+*.ttf
+*.unitypackage
+*.wav
+*.xlsx
+*.zip
+
+# scripts
+*.cs
+*.h
+*.m
+*.mm
+
+# miscellaneous
+*.cginc
+*.xml
+*.txt
+*.swatch
+*.md
+*.json
+*.mdb
+*.bat
+*.sh
+*.jar
+*.keystore
+*.url
+*.html
+*.md5
+*.plist
+*.modulemap
+*.entitlements
+*.framework
+*.properties
+*.XML
+*.shader
+*.compute
+
+# all meta except which contains GUID
+*.meta
+
+!*.FBX.meta
+!*.fbx.meta
+!*.otf.meta
+!*.shader.meta
+!*.ttf.meta
