+++ Editor/ignore.txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b5449731d8719ce459839efa75103af3
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
