+++ LICENSE.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b6d9371dcdcc09f448196d6ff28c312c
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
