+++ README.md.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c03ec6755814d544d873d404ad3afec4
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
