+++ package.json
@@ -1,0 +1,12 @@
+{
+  "name": "com.github.idbrii.unity-findrefs",
+  "displayName": "findrefs",
+  "version": "2.0.0",
+  "unity": "2019.4",
+  "description": "Find references in your project",
+  "keywords": [ "editor", "reference", "search" ],
+  "category": "productivity",
+  "dependencies": {},
+  "samples": [
+  ]
+}
