+++ package.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6fa60b0e4252a1d41ba841d6f0f1d0d0
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
